requirements.txt
pypdf
langchain
python-dotenv
faiss-cpu
streamlit
transformers
huggingface_hub
langchain_community